fix: allow fractional delay_hours in campaign steps validation
Wait nodes use fractional hours (e.g. 0.666 for 40 mins) but the
Zod schema required integers. Removed .int() constraint.

Build verified locally.

+++ src/lib/validations.ts
@@ -211,7 +211,7 @@ export const createCampaignSchema = z.object({
     step_type: z.enum(["send", "condition"]).default("send"),
     wa_template_name: z.string(),
     template_id: z.string().optional().or(z.literal("")),
-    delay_hours: z.number().int().min(0),
+    delay_hours: z.number().min(0),
     wa_template_params: z.array(z.string()),
     wa_template_param_names: z.array(z.string()).optional().default([]),
     condition: z.object({
